Correccion entero a decimal de monedas en retenciones RH

+++ Convertidor/Data/Entities/Rh/RH_H_RETENCIONES_FAOV.cs
@@ -9,11 +9,11 @@
         public string NOMBRES_APELLIDOS { get; set; }
         public string DESCRIPCION_CARGO { get; set; }
         public DateTime FECHA_INGRESO { get; set; }
-        public int MONTO_FAOV_TRABAJADOR { get; set; }
+        public decimal MONTO_FAOV_TRABAJADOR { get; set; }
 
-        public int MONTO_FAOV_PATRONO { get; set; }
+        public decimal MONTO_FAOV_PATRONO { get; set; }
 
-        public int MONTO_TOTAL_RETENCION { get; set; }
+        public decimal MONTO_TOTAL_RETENCION { get; set; }
 
         public string FECHA_NOMINA { get; set; }
