+++ Convertidor/Dtos/Rh/RhTmpRetencionesFaovDto.cs
@@ -10,9 +10,9 @@
         public string NombresApellidos { get; set; }
         public string DescripcionCargo { get; set; }
         public DateTime FechaIngreso { get; set; }
-        public int MontoFaovTrabajador { get; set; }
-        public int MontoFaovPatrono { get; set; }
-        public int MontoTotalRetencion { get; set; }
+        public decimal MontoFaovTrabajador { get; set; }
+        public decimal MontoFaovPatrono { get; set; }
+        public decimal MontoTotalRetencion { get; set; }
         public string FechaNomina { get; set; }
         public string SiglasTipoNomina { get; set; }
         public string RegistroConcat { get; set; }
